# docs/openapi.yaml
openapi: 3.1.0
info:
  title: E-Ticketing Transportasi Publik API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /auth/refresh:
    post:
      tags:
        - Otentikasi
      summary: Refresh token JWT
      description: Hasilkan token akses baru menggunakan refresh token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            examples:
              refresh_token:
                summary: Contoh refresh token
                value:
                  refresh_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Token berhasil di-refresh
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/RefreshTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Masukkan token JWT yang diperoleh dari endpoint login admin
  schemas:
...
    RefreshTokenRequest:
      type: object
      properties:
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      required:
        - refresh_token
    RefreshTokenResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        expires_at:
          type: integer
          format: int64
          example: 1727600000
      required:
        - token
        - expires_at
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Badan permintaan tidak valid"
      required:
        - error
  responses:
    BadRequestError:
      description: Permintaan buruk - Data input tidak valid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthorizedError:
      description: Tidak sah - Token JWT tidak valid atau hilang
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...